refactor: split window.py into per-concern mixins
window.py was 1977 lines and 104 methods in one class -- 35% of the codebase
and the only large file with no test coverage. It is now 286 lines: the
template, __init__, _load_mail_view and the window lifecycle. Everything else
moved verbatim into seven mixins, none over 400 lines:

  window_accounts.py  178  the switcher, adding accounts, opening the composer
  window_actions.py   363  read/star, the flag worker, the row context menu
  window_move.py      317  archive/trash/move with the undo window
  window_folders.py   181  the sidebar tree, its rows and selection
  window_list.py      213  the conversation list, search, scroll paging
  window_reader.py    233  the reading pane, message bodies, attachments
  window_sync.py      395  syncing, the Outbox, the connection banner

Bodies were sliced by line range, so the diff is a move rather than a rewrite.

Two supporting modules. window_types.py holds the constants and records the
modules share, separate from window.py so the mixins can import them without
importing the window that imports them. window_parts.py declares the shared
surface once -- the template widgets, the window state, and the 45 methods the
modules call on each other -- so each mixin type-checks alone instead of every
file redeclaring the lot. Its stubs raise NotImplementedError, and a check
confirms all 45 resolve to real implementations rather than the stub.

Two things the plan did not anticipate:

Adw.ApplicationWindow is now listed *last* in the bases. The mixins share
MainWindowParts, which is the window type when type-checking, and C3 requires a
base to follow its own subclasses. Position does not change the GType parent, so
Gtk.Template still sees AdwApplicationWindow -- verified.

MainWindowParts only inherits the window type under TYPE_CHECKING. It has to be
a plain object at runtime: Gtk.Template checks that the template's declared
parent matches the instance's direct parent GType, and a mixin that were itself
a GObject subclass makes the direct parent a mixin, which makes the template
refuse to build. Both alternatives were tried and rejected on evidence.

Splitting also exposed three complexity findings that window.py's blanket
per-file-ignore had been hiding, all now fixed rather than re-suppressed:
_refresh_conversations (60 lines, complexity 11, 14 branches) is five named
steps; _flag_worker and _body_worker take a FlagChange/BodyRequest snapshot
instead of 7 and 6 loose arguments -- frozen, which is what the threading model
wants anyway. window.py's E402/C901/PLR0912/PLR0913 exemptions are gone.

All nine modules are registered in src/meson.build, and the five with
translatable strings in po/POTFILES.in; window.py has none left, so its entry
is removed.

Verified: ruff and pyright clean, 183 tests pass, the Flatpak builds, and the
app runs against the real database with no template or MRO error.
