+++ po/POTFILES.in
@@ -8,7 +8,11 @@ src/postcard/application.py
 src/postcard/composer_window.py
 src/postcard/message_view.py
 src/postcard/preferences_dialog.py
-src/postcard/window.py
+src/postcard/window_accounts.py
+src/postcard/window_actions.py
+src/postcard/window_move.py
+src/postcard/window_reader.py
+src/postcard/window_sync.py
 src/ui/account-dialog.blp
 src/ui/accounts-dialog.blp
 src/ui/composer-window.blp
