+++ pyproject.toml
@@ -38,8 +38,8 @@ max-args = 5
 [tool.ruff.lint.per-file-ignores]
 # gi.require_version() must run before the gi.repository import it configures,
 # which pushes that import (and everything after it) below E402's threshold.
-# (window.py, composer_window.py and message_view.py also need E402 -- their
-# entries live in the deferred block below so each file has one key.)
+# (composer_window.py and message_view.py need it too; their entries are below,
+# so each file keeps a single key.)
 "src/postcard/application.py" = ["E402"]
 "src/postcard/core/avatars.py" = ["E402"]
 "src/postcard/core/secrets.py" = ["E402"]
@@ -57,15 +57,16 @@ max-args = 5
 # a constructor parameter. Mitigated by making them keyword-only instead.
 "src/postcard/core/models/*.py" = ["PLR0913"]
 
-# --- Deferred: delete each entry as the refactor phase named clears it -------
+# --- Remaining carve-outs ----------------------------------------------------
 # save_account and save_email take the account/email columns as parameters; both
 # insert into a table wider than 5 columns and have no natural carrier object
 # (unlike save_incoming_email, which now takes a MessageHeader).
 "src/postcard/core/store/database.py" = ["PLR0913"]
 # fetch_mailbox and set_flag take the connection details plus what to act on.
 "src/postcard/mail_sync.py" = ["PLR0913"]
-# Phase 7 (splitting window.py and the long __init__s): oversized functions.
-"src/postcard/window.py" = ["E402", "C901", "PLR0912", "PLR0913"]
+# composer_window and message_view have gi.require_version gates and one wide
+# constructor each; core/compose.build_mime_message takes the six parts of a
+# message. threader.group's complexity is the union-find itself.
 "src/postcard/composer_window.py" = ["E402", "PLR0913"]
 "src/postcard/message_view.py" = ["E402", "PLR0913"]
 "src/postcard/core/compose.py" = ["PLR0913"]
