+++ src/meson.build
@@ -35,6 +35,15 @@ postcard_sources = [
   'postcard/main.py',
   'postcard/application.py',
   'postcard/window.py',
+  'postcard/window_types.py',
+  'postcard/window_parts.py',
+  'postcard/window_accounts.py',
+  'postcard/window_actions.py',
+  'postcard/window_move.py',
+  'postcard/window_folders.py',
+  'postcard/window_list.py',
+  'postcard/window_reader.py',
+  'postcard/window_sync.py',
   'postcard/conversation_row.py',
   'postcard/avatar_loader.py',
   'postcard/message_view.py',
